refactor: extract exports analyzer to separate class
