+++ lib/fileAnalyzer.ts
@@ -1,11 +1,8 @@
-import Debug from 'debug';
 import fs from 'fs-extra';
-import parser from 'solidity-parser-antlr';
 import stripComments from 'strip-json-comments';
+import { ExportsAnalyzer, ExportsAnalyzerResult } from './exportsAnalyzer';
 import { RegistredImport } from './importRegistry';
 import { ImportsAnalyzerResult, ImportsAnalyzer } from './importsAnalyzer';
-
-const error = Debug('sol-merger:error');
 
 export class FileAnalyzer {
   filename: string;
@@ -15,7 +12,7 @@ export class FileAnalyzer {
    */
   static buildExportBody(
     analyzedFile: FileAnalyzerResult,
-    e: FileAnalyzerExportsResult,
+    e: ExportsAnalyzerResult,
     newName: string,
     globalRenames: RegistredImport[],
   ): string {
@@ -59,7 +56,8 @@ export class FileAnalyzer {
     const importsAnalyzer = new ImportsAnalyzer(contents);
     const imports = importsAnalyzer.analyzeImports();
 
-    const exports = this.analyzeExports(contents);
+    const exportsAnalyzer = new ExportsAnalyzer(contents);
+    const exports = exportsAnalyzer.analyzeExports();
     return {
       filename: this.filename,
       contents,
@@ -67,55 +65,11 @@ export class FileAnalyzer {
       exports,
     };
   }
-
-  /**
-   * Analyzes all the exports of the file (Contract, Interface, Library)
-   *
-   * Single export statement to process. Basicly it analizes next things:
-   *
-   * 1. Get the type of the export
-   * 2. Get the body of the export
-   * 3. Get inheritance of the specifier
-   *
-   */
-  analyzeExports(contents: string): FileAnalyzerExportsResult[] {
-    try {
-      const ast = parser.parse(contents, { loc: true, range: true });
-      const results: FileAnalyzerExportsResult[] = [];
-      const exportRegex = /(contract|library|interface)\s+([a-zA-Z_$][a-zA-Z_$0-9]*)\s*([\s\S]*?)\{/;
-      parser.visit(ast, {
-        ContractDefinition: (node) => {
-          const contract = contents.substring(node.range[0], node.range[1] + 1);
-          const group = exportRegex.exec(contract);
-          const [match, _, __, is] = group;
-          results.push({
-            is: is,
-            name: node.name,
-            type: node.kind as any,
-            body: contract.substring(match.length - 1),
-          });
-        },
-      });
-      return results;
-    } catch (e) {
-      if (e instanceof (parser as any).ParserError) {
-        error(e.errors);
-      }
-      return [];
-    }
-  }
 }
 
 export interface FileAnalyzerResult {
   filename: string;
   contents: string;
   imports: ImportsAnalyzerResult[];
-  exports: FileAnalyzerExportsResult[];
-}
-
-export interface FileAnalyzerExportsResult {
-  type: 'contract' | 'library' | 'interface';
-  name: string;
-  is: string;
-  body: string;
+  exports: ExportsAnalyzerResult[];
 }
